refactor(avro): inline Utils into AvroExtension
Fold the avro `Utils` object (emitAvroSchema / isEnum / flatten) into
AvroExtension.kt as private top-level functions, since it was only used
by AvroExtension. Removes the separate Utils.kt file.
